feat: se corrigen errores de codigo

+++ src/main/java/com/brayanbautista/kinalapp/repository/DetalleVentaRepository.java
@@ -4,6 +4,6 @@ import com.brayanbautista.kinalapp.entity.DetalleVenta;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface DetalleVentaRepository extends JpaRepository<DetalleVenta, Integer> {
-    List<DetalleVenta> findByVenta_CodigoVenta(Integer codigoVenta);
+public interface DetalleVentaRepository extends JpaRepository<DetalleVenta, Long> {
+    List<DetalleVenta> findByVenta_CodigoVenta(Long codigoVenta);
 }
